+++ threedi_model_migration/tests/conftest.py
@@ -38,6 +38,4 @@ def repository(tmp_path_factory):
     con.close()
     hg.add(repo_path, "db2.sqlite")
     hg.commit(repo_path, "db2.sqlite", "My second commit")
-    return Repository(
-        base_path=tmp_path, slug="testrepo", remote="https://non.existing"
-    )
+    return Repository(base_path=tmp_path, slug="testrepo")